requirements.txt
discord.py[voice]
lxml
psutil
asyncpg>=0.12.0
click
parsedatetime
lru_dict
python-dateutil
config
bs4
pymongo